% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{dat}
\alias{dat}
\title{Example dataset: dat}
\format{
A data frame with 10 rows and 5 columns:
\describe{
\item{suid}{Subject ID's}
\item{total}{Total number of trials}
\item{stage}{Binary predictor variable (0/1)}
\item{M1}{Count data for Marker 1}
\item{M2}{Count data for Marker 2}
}
}
\source{
Imported from CSV and saved as RData
}
\usage{
dat
}
\description{
A sample dataset used for demonstrating the function.
}
\examples{
data(dat)
head(dat)
}
\keyword{datasets}
